refactor(node-compat): split capability checking into abstractions
