+++ packages/node-compat/src/internal/index.ts
@@ -1,0 +1,7 @@
+export { collectRequiredCapabilities } from "../collect.js";
+export { checkCapabilities } from "../check.js";
+export {
+  CAPABILITY_KEYS,
+  CAPABILITY_MATRIX,
+  CapabilityStatus,
+} from "../matrix.js";
